Added functions used to score
Created a click function for the Score button on the scoring page; when clicked, the function will read the user's inputs for each piece and calculate a total score based on this information. The function is called twice, once for each side. Then, based on the final score the display is updated to show the winner, if there is one.